source/backend/requirements.txt
# fastapi
fastapi==0.89.1
uvicorn==0.20.0
email-validator==1.3.0

# data
sqlalchemy[asyncio]==2.0.0rc2
asyncpg==0.27.0
msgspec==0.12.0

# communication with the Gateway
grpcio==1.51.1
protobuf==4.21.12

# security
python-dotenv==0.21.0
argon2-cffi==21.3.0

# server
gunicorn==20.1.0
uvloop==0.17.0; platform_system!="Windows" # MagicStack/uvloop#14
